Create box_shadow_transparent
Recently, I had a project that required using css3 based shadows on top of images with varying backgrounds. The shadow needed to be transparent. The standard box shadow did not have this. So, I used RGBa transparency here. The setup I used for a basic black drop shadow with 50% transparency.

-moz-box-shadow: 2px 2px 2px rgba(0,0,0,0.5);
 -webkit-box-shadow: 2px 2px 2px rgba(0,0,0,0.5);
 box-shadow: 2px 2px 2px rgba(0,0,0,0.5);

The box-shadow atrributes are (in order):
Horizontal Offset – Positive values move the shadow the right, negative values to the left.
Vertical Offset – Positive values move the shadow down, negative moves it up.
Blur – Can also be thought of as the “sharpness” of the shadow. 0px would make a very sharp shadow, higher values will increase the blur.
Color - Can be any valid hex or rgb color, but in this case we’re using RGBa which adds an alpha channel to achieve transparency.